Refactor provider registry loader
